requirements.txt
requests==2.32.3
supabase==2.15.0
python-dotenv==1.0.1